feat: move daily forecast to screen state feat: use localization and theme in daily forecast
