+++ lib/l10n/app_en.arb
@@ -6,5 +6,6 @@
   "rain": "Rain",
   "humidity": "Humidity",
   "wind": "Wind",
-  "today": "Today"
+  "today": "Today",
+  "daily_forecast_title": "Next 7 days"
 }
