+++ lib/l10n/app_localizations_ar.dart
@@ -28,4 +28,7 @@ class AppLocalizationsAr extends AppLocalizations {
 
   @override
   String get today => 'اليوم';
+
+  @override
+  String get daily_forecast_title => 'سبع ايام القادمين';
 }
