+++ lib/presentation/screen/component/hourly_forecast_section.dart
@@ -30,7 +30,7 @@ class HourlyForecastSection extends StatelessWidget {
             ),
           ),
         ),
-        const SizedBox(height: 24),
+        const SizedBox(height: 12),
         SizedBox(
           height: 140,
           child: ListView.builder(
