+++ lib/presentation/screen/weather_screen_state.dart
@@ -1,3 +1,5 @@
+import '../../gen/assets.gen.dart';
+
 class HourlyForecastUiState {
   final String forecastImage;
   final String temperatureDegree;
@@ -7,6 +9,20 @@ class HourlyForecastUiState {
     required this.forecastImage,
     required this.temperatureDegree,
     required this.hour,
+  });
+}
+
+class DailyForecastUiState {
+  final String day;
+  final AssetGenImage forecastImage;
+  final int lowTemperature;
+  final int highTemperature;
+
+  const DailyForecastUiState({
+    required this.day,
+    required this.forecastImage,
+    required this.lowTemperature,
+    required this.highTemperature,
   });
 }
 
@@ -45,5 +61,50 @@ List<HourlyForecastUiState> dummyHourlyForecast = [
     forecastImage: 'assets/images/day_clear_sky.png',
     temperatureDegree: '24°C',
     hour: '15:00',
+  ),
+];
+
+List<DailyForecastUiState> dummyDailyForecast = [
+  DailyForecastUiState(
+    day: 'Thursday',
+    forecastImage: Assets.images.dayClearSky,
+    lowTemperature: 20,
+    highTemperature: 32,
+  ),
+  DailyForecastUiState(
+    day: 'Friday',
+    forecastImage: Assets.images.dayOvercast,
+    lowTemperature: 20,
+    highTemperature: 32,
+  ),
+  DailyForecastUiState(
+    day: 'Saturday',
+    forecastImage: Assets.images.dayDrizzleModerate,
+    lowTemperature: 19,
+    highTemperature: 31,
+  ),
+  DailyForecastUiState(
+    day: 'Sunday',
+    forecastImage: Assets.images.dayDrizzleIntensity,
+    lowTemperature: 18,
+    highTemperature: 30,
+  ),
+  DailyForecastUiState(
+    day: 'Monday',
+    forecastImage: Assets.images.dayClearSky,
+    lowTemperature: 20,
+    highTemperature: 32,
+  ),
+  DailyForecastUiState(
+    day: 'Tuesday',
+    forecastImage: Assets.images.dayDrizzleModerate,
+    lowTemperature: 20,
+    highTemperature: 32,
+  ),
+  DailyForecastUiState(
+    day: 'Wednesday',
+    forecastImage: Assets.images.dayDepositingRimeFog,
+    lowTemperature: 21,
+    highTemperature: 33,
   ),
 ];
